+++ .gitignore
@@ -60,5 +60,6 @@ typings/
 # next.js build output
 .next
 
-#phea config
+#phea local testing
 index.js
+index.*.js
