+++ examples/multi-light-example.js
