--- src/textures/texture.js
@@ -1,7 +1,0 @@
-const sine = require('./sine');
-const square = require('./square');
-
-module.exports =  {
-    'sine': sine,
-    'square': square
-}
